+++ macros/DramaticTasks.js
@@ -1,12 +1,16 @@
-const version = 'v1.1';
+const version = 'v1.2';
 const sm = game.modules.get('swademacros')?.api.sm;
+
 const chatimage = 'icons/commodities/tech/detonator-timer.webp';
-const coreRulesLink = '@Compendium[swade-core-rules.swade-rules.G9H6LoPoQu10TI4R]{Dramatic Tasks}';
-let coreRules = sm.isModuleOn("swade-core-rules");
+const challengeTrackerFlag = sm.isModuleOn("challenge-tracker");
 
 main();
 
 function main() {  
+  let challengeTrackerHTML='';
+  if (challengeTrackerFlag) {
+    challengeTrackerHTML = 'checked';
+  }
   let template =`
 <style type="text/css">
 .tg  {border-collapse:collapse;border-spacing:0;}
@@ -22,18 +26,20 @@ function main() {
 <tbody>
   <tr>
     <td class="tg-xwyw">
-        <h3><b style="font-size:20px">Players</b></h3>
-        <select id="playersAmount" name="playersAmount">
-          <option value="1" selected="selected">1</option>
-          <option value="2">2</option>
-          <option value="3">3</option>
-          <option value="4">4</option>
-          <option value="5">5</option>
-          <option value="6">6</option>
-        </select> 
+        <b>Players</b>
+        <p>
+          <select id="playersAmount" name="playersAmount">
+            <option value="1" selected="selected">1</option>
+            <option value="2">2</option>
+            <option value="3">3</option>
+            <option value="4">4</option>
+            <option value="5">5</option>
+            <option value="6">6</option>
+          </select> 
+        </p>
     </td>
     <td class="tg-0lax">
-      <h3><b style="font-size:20px">Difficult</b></h3>
+      <b>Difficult</b>
       <p>
       <select id="difficult" name="difficult">
         <option value="challenging" selected="selected">Challenging</option>
@@ -43,6 +49,44 @@ function main() {
       </p>
     </td>
   </tr>
+
+  <tr>
+    <td class="tg-xwyw" colspan="2">
+      <h3>Custom</h3>
+      <p>Only change below if you want to customize.</p>
+    </td>
+  </tr>
+
+  <tr>
+    <td class="tg-xwyw">
+        <b>Tokens</b>
+        <p>
+          <input id="customTokens" type="number" min="1" max="100" value="0">    
+        </p>
+    </td>
+    <td class="tg-0lax">
+      <b>Turns</b>
+      <p>
+        <input id="customTurns" type="number" min="1" max="100" value="0">    
+      </p>
+    </td>
+  </tr>
+
+  <tr>
+    <td class="tg-xwyw" colspan="2">
+      <h3><b>Options</b></h3>
+    </td>
+  </tr>
+  
+  <tr>
+    <td class="tg-xwyw">
+      <b>Challenge Tracker</b>
+    </td>
+    <td class="tg-xwyw">
+      <input id="challengeTracker" type="checkbox" ${challengeTrackerHTML}>
+    </td>    
+  </tr>
+  
 </tbody>
 </table>
 `;
@@ -69,15 +113,18 @@ function main() {
 async function dramaticTask(html) {
   const players = html.find("#playersAmount")[0].value;    
   const difficult = html.find("#difficult")[0].value;      
-  let message=``;
-    
-  if (coreRules === true) {
-    message = `<div class="swade-core"><h2><img style="vertical-align:middle" src=${chatimage} width="28" height="28"> ${coreRulesLink}</h2></div>`;
-  }  else {
-    message = `<h2><img style="vertical-align:middle" src=${chatimage} width="28" height="28"> Dramatic Tasks</h2>`;      
-  }
+  const customTokens = html.find("#customTokens")[0].value;    
+  const customTurns = html.find("#customTurns")[0].value;      
+  const challengeTracker = html.find("#challengeTracker")[0].value;      
   
-  message += calculateTaskTokens(players, difficult);
+  let message=``;
+  message = `<h2><img style="vertical-align:middle" src=${chatimage} width="28" height="28"> Dramatic Tasks</h2>`;      
+
+  if ( customTokens!=0 && customTurns!=0 ) {
+    message += customDramaticTask(customTokens, customTurns, challengeTracker);
+  } else {
+    message += calculateTaskTokens(players, difficult, challengeTracker);
+  }
 
   // send message
   let chatData = {
@@ -87,7 +134,7 @@ async function dramaticTask(html) {
   ChatMessage.create(chatData, {}); 
 }
 
-function calculateTaskTokens(players, difficult) {
+function calculateTaskTokens(players, difficult, challengeTracker=false) {
   let tasksTokens=0;
   let tasksTurns=0;
   let difficultName='';
@@ -107,57 +154,22 @@ function calculateTaskTokens(players, difficult) {
   }
   message+=`<p>This is <b>${difficult}</b> for <b>${players}</b> player(s).</p>`;
   message+=`<ul><li>Task Tokens: <b style="color:red;">${tasksTokens}</b></li><li>Rounds: <b style="color:red;">${tasksTurns}</b></p>`;
+  
+  if (challengeTracker) {
+    ChallengeTracker.open(tasksTokens, tasksTurns);
+  }  
+  
   return message;
 }
 
-/*
-// prob ===============================================
-async function resolveOneTask(skills) {    
-    let outcome = 0; // "outcome" is "task tokens", ie. one per success plus one per raise
-    for (const skillDie of skills) { 
-        const diceFormula ="{" + skillDie + "x,d6x}kh";
-        let diceRoll = new Roll(diceFormula);
-        diceRoll.evaluate({async: false});
-        outcome += Math.floor(diceRoll.result / 4);
-        //console.log(`Result ${diceRoll.result} outcome is now ${outcome}`);
-    }
-    return outcome;
+function customDramaticTask(tokens, turns, challengeTracker=false) {
+  let message=``;
+  message+=`<ul><li>Task Tokens: <b style="color:red;">${tokens}</b></li><li>Rounds: <b style="color:red;">${turns}</b></p>`;
+  
+  if (challengeTracker) {
+    ChallengeTracker.open(tokens, turns);
+  }
+  
+  return message;
 }
 
-var attemptedTasks = Array.from(Array(1000).keys());
-var sucessfulTasks = attemptedTasks.reduce((acumulador, numero) => {
-  let outcome = resolveOneTask(["d8", "d8", "d8", "d8"]);
-  if (outcome >= 6) { 
-    acumulador++;
-  }
-  return acumulador;
-}, 0)
-
-let fraction = sucessfulTasks / attemptedTasks.length;
-console.log(`Did ${attemptedTasks.length} tasks, ${sucessfulTasks} succeeded (${fraction})`);
-
-javier
-async function resolveOneTask(skills) {    
-    let outcome = 0; // "outcome" is "task tokens", ie. one per success plus one per raise
-    for (const skillDie of skills) { 
-        const diceFormula ="{" + skillDie + "x,d6x}kh";
-        let diceRoll = new Roll(diceFormula);
-        diceRoll.evaluate({async: false});
-        outcome += Math.floor(diceRoll.result / 4);
-        //console.log(`Result ${diceRoll.result} outcome is now ${outcome}`);
-    }
-    return outcome;
-}
-
-var attemptedTasks = Array.from(Array(1000).keys());
-var sucessfulTasks = attemptedTasks.reduce( async (acumulador) => {
-  let outcome = resolveOneTask(["d8", "d8", "d8", "d8"]);
-  if (outcome >= 6) { 
-    await accumulador;
-    return acumulador++;
-  }
-}, 0)
-
-let fraction = sucessfulTasks / attemptedTasks.length;
-console.log(`Did ${attemptedTasks.length} tasks, ${sucessfulTasks} succeeded (${fraction})`);
-*/